Refactoring: Separate logic with voice messages in another service
